% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runSGD_cv.R
\name{calculateSGD_cv}
\alias{calculateSGD_cv}
\alias{runSGD_cv}
\alias{calculateSGD_cv,ANY-method}
\alias{calculateSGD_cv,SummarizedExperiment-method}
\alias{calculateSGD_cv,SingleCellExperiment-method}
\alias{runSGD_cv,SummarizedExperiment-method}
\alias{runSGD_cv,SingleCellExperiment-method}
\title{Perform a stochastic gradient descent generalized matrix factorization (sgdGMF) on cells,
based on the expression or mass spectrometry data in a SingleCellExperiment object.}
\usage{
calculateSGD_cv(x, ...)

runSGD_cv(x, ...)

\S4method{calculateSGD_cv}{ANY}(
  x,
  family = poisson(),
  ncomponents = 1:10,
  ntop = NULL,
  X = NULL,
  Z = NULL,
  offset = NULL,
  weights = NULL,
  subset_row = NULL,
  scale = FALSE,
  transposed = FALSE,
  BSPARAM = bsparam(),
  BPPARAM = SerialParam(),
  control.init = list(),
  control.alg = list(),
  control.cv = list(),
  penalty = list(),
  method = "sgd",
  sampling = "block"
)

\S4method{calculateSGD_cv}{SummarizedExperiment}(
  x,
  ...,
  exprs_values = "counts",
  assay.type = exprs_values,
  family = poisson()
)

\S4method{calculateSGD_cv}{SingleCellExperiment}(
  x,
  ...,
  exprs_values = "counts",
  dimred = NULL,
  n_dimred = NULL,
  assay.type = exprs_values,
  family = poisson()
)

\S4method{runSGD_cv}{SummarizedExperiment}(x, ..., name = "cv_SGD")

\S4method{runSGD_cv}{SingleCellExperiment}(x, ..., altexp = NULL, name = "cv_SGD")
}
\arguments{
\item{x}{For \code{calculateSGD}, a numeric matrix of expression counts or mass spectrometry intensities where
rows are features and columns are cells.
Alternatively, a \linkS4class{SummarizedExperiment} or \linkS4class{SingleCellExperiment} containing such a matrix.
For \code{runSGD}, a \linkS4class{SingleCellExperiment} object containing such a matrix.}

\item{...}{For the \code{calculateSGD} generic, additional arguments to pass to specific methods.
For the SummarizedExperiment and SingleCellExperiment methods, additional arguments to pass to the ANY method.

For \code{runSGD}, additional arguments to pass to \code{calculateSGD}.}

\item{family}{The distribution family that is used for the estimation of the parameters.}

\item{ncomponents}{Numeric vector indicating the different number of components used in cross-validation.}

\item{ntop}{Numeric scalar specifying the number of features with the highest variances to use for dimensionality reduction.
Default uses all features.}

\item{X}{Sample-level covariate matrix. Defaults to column of ones.}

\item{Z}{Feature-level covariate matrix. Defaults to column of ones.}

\item{offset}{offset matrix with same dimensions as x that is added to the linear predictor. Note that if family = poisson(), this should therefore be on the log-scale}

\item{weights}{weight matrix with same dimensions as x that determines the weight of each observation.}

\item{subset_row}{Vector specifying the subset of features to use for dimensionality reduction.
This can be a character vector of row names, an integer vector of row indices or a logical vector.}

\item{scale}{Logical scalar, should the expression values be standardized? Not recommended for non-Gaussian data.}

\item{transposed}{Logical scalar, is \code{x} transposed with cells in rows?}

\item{BSPARAM}{A \linkS4class{BiocSingularParam} object specifying which algorithm should be used to perform the PCA.
This is used in \code{runPCA} to put all information in the sample latent factors.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object specifying whether the cross-validation
should be parallelized. If BPPARAM$workers > 1 and control.cv$parallel and control.cv$nthreads are
not specified, parallelization is enabled with nthreads = BPPARAM$workers.}

\item{control.init}{control parameters for the initialization, used in the \code{sgdGMF} package. See \link{sgdgmf.init} and \link{set.control.init}.}

\item{control.alg}{control parameters for the estimation, used in the \code{sgdGMF} package. See \link{sgdgmf.fit} and \link{set.control.alg}.}

\item{control.cv}{control parameters for the cross-validation, used in the \code{sgdGMF} package. See \link{sgdgmf.cv} and \link{set.control.cv}.}

\item{penalty}{ridge penalty added for the estimation of the parameters in the \code{sgdGMF} package. see \link{sgdgmf.fit}.}

\item{method}{estimation algorithm from the \code{sgdGMF} package used. see \link{sgdgmf.fit}.}

\item{sampling}{sub-sampling strategy to use if method = "sgd". See \link{sgdgmf.fit} from the \code{sgdGMF} package.}

\item{exprs_values}{Alias to \code{assay.type}.}

\item{assay.type}{Integer scalar or string indicating which assay of \code{x} contains the values of interest.}

\item{dimred}{String or integer scalar specifying the existing dimensionality reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if \code{dimred} is specified.}

\item{name}{String specifying the name to be used to store the result in the \code{\link{reducedDims}} of the output.}

\item{altexp}{String or integer scalar specifying an alternative experiment containing the input data.}
}
\value{
A table containing the summary statistics of the cross-validation. If a
\linkS4class{SummarizedExperiment} or \linkS4class{SingleCellExperiment} was
given as input, this is stored in the metadata of this object.
}
\description{
Perform a stochastic gradient descent generalized matrix factorization (sgdGMF) on cells,
based on the expression or mass spectrometry data in a SingleCellExperiment object.
}
\details{
sgdGMF uses sampling of the data to estimate the parameters, which can alter with different seeds. Also, cross-validation
puts a random selection of values to missing. This means that the result will change slightly across different runs.
For full reproducibility, users should call \code{\link{set.seed}} prior to running \code{runSGD} with such algorithms.
(Note that this includes \code{BSPARAM=\link{bsparam}()}, which uses approximate algorithms by default.)
}
\section{Feature selection}{

This section is relevant if \code{x} is a numeric matrix with features in rows and cells in columns;
or if \code{x} is a \linkS4class{SingleCellExperiment} and \code{dimred=NULL}.
In the latter, the expression values are obtained from the assay specified by \code{assay.type}.

The \code{subset_row} argument specifies the features to use for dimensionality reduction.
The aim is to allow users to specify highly variable features to improve the signal/noise ratio,
or to specify genes in a pathway of interest to focus on particular aspects of heterogeneity.

If \code{subset_row=NULL}, the \code{ntop} features with the largest variances are used instead.
We literally compute the variances from the expression values without considering any mean-variance trend,
so often a more considered choice of genes is possible, e.g., with \pkg{scran} functions.
Note that the value of \code{ntop} is ignored if \code{subset_row} is specified.

If \code{scale=TRUE}, the expression values for each feature are standardized so that their variance is unity.
This will also remove features with standard deviations below 1e-8.
}

\examples{
example_sce <- mockSCE()
example_sce <- runSGD_cv(example_sce, exprs_values="counts", family = poisson(), ncomponents = c(1:5))
head(metadata(example_sce)[["SGD"]])
example_sce <- runSGD(example_sce, exprs_values="counts", family = poisson(), ncomponents = 3)
reducedDimNames(example_sce)
head(reducedDim(example_sce))
}
\seealso{
\code{\link[sgdGMF]{sgdgmf.cv}}, for the underlying calculations.
}
\author{
Alexandre Segers
}
