% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SGDImpute.R
\name{SGDImpute}
\alias{SGDImpute}
\alias{SGDImpute,ANY-method}
\alias{SGDImpute,SummarizedExperiment-method}
\alias{SGDImpute,SingleCellExperiment-method}
\title{Impute missing values based on the results of runSGD.}
\usage{
SGDImpute(x, ...)

\S4method{SGDImpute}{ANY}(x, sgdGMF_reducedDims)

\S4method{SGDImpute}{SummarizedExperiment}(x, sgdGMF_reducedDims, exprs_values = 1, assay.type = exprs_values)

\S4method{SGDImpute}{SingleCellExperiment}(
  x,
  reducedDimName = "SGD",
  sgdGMF_reducedDims = reducedDim(x, reducedDimName),
  ...,
  exprs_values = 1,
  assay.type = exprs_values
)
}
\arguments{
\item{x}{a numeric matrix of expression counts or mass spectrometry intensities containing missing values and with
features in the rows and samples in columns.
Alternatively, a \linkS4class{SummarizedExperiment} or \linkS4class{SingleCellExperiment} containing such a matrix.}

\item{sgdGMF_reducedDims}{the output obtained by \code{'runSGD'} or \code{'calculateSGD'}. If \code{x} is a
\linkS4class{SingleCellExperiment}, \code{sgdGMF_reducedDims} is taken from \code{\link{reducedDim}(x, 'SGD')}.}

\item{exprs_values}{Alias to \code{assay.type}.}

\item{assay.type}{Integer scalar or string indicating which assay of \code{x} contains the values of interest.}
}
\value{
A SingleCellExperiment object is returned with an extra assay containing the imputed values.
}
\description{
Impute missing values based on the results of runSGD.
}
\details{
Imputation is only possible after running \code{'runSGD'} using all features.
Therefore, \code{subset_row} or \code{ntop} should be set to NULL.
}
\examples{
example_sce <- mockSCE()


example_sce <- logNormCounts(example_sce)

example_sce <- runSGD(example_sce)
reducedDimNames(example_sce)
head(reducedDim(example_sce))
}
\seealso{
\code{\link[BiocSGDGMF]{runSGD}}, to conveniently obtain the matrix factorization.
\code{\link[BiocSGDGMF]{plotSGD}}, to conveniently visualize the results.
}
\author{
Alexandre Segers
}
